Add fuzzer for model/converter/thrift/zipkin
This fuzzes the Thrift deserializer. Admittedly this primarily tests
the thrift parser, which is a 3rd party dependency. However, it
doesn't hurt to make sure the Thrift parser is behaving well and not
rely on the dependency to do its own fuzzing.

In particular, after only several minutes of running on random input,
we already manage to find inputs that cause deserialisation to take
upward of 40 seconds. It also shows extreme memory usage.

This might be worth investigating further.

Signed-off-by: Dominik Honnef <dominik@honnef.co>